# docs/schemas/openapi/pangu2-api-v1.yaml
openapi: "3.1.0"
info:
  title: PANGU2 API
  description: |
    PANGU2 前后端共享 API 契约。
    Admin control panel and user DApp share this schema.
    Backend MUST implement; frontend MUST consume via generated client.
    No duplicate DTOs or business formulas in frontend.
  version: "1.0.0"
  contact:
    name: PANGU2 Team
servers:
  - url: http://localhost:8080/api/v1
    description: LOCAL development
  - url: https://testnet.pangu2.io/api/v1
    description: BSC TESTNET
...
paths:
...
  /projects/pangu2/auth/verify:
    post:
      tags: [Auth]
      operationId: verifySignature
      summary: Verify wallet signature and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_address, signature]
              properties:
                wallet_address:
                  $ref: "#/components/schemas/EvmAddress"
                signature:
                  type: string
                  description: EIP-191 signed message
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope_Session"
...
components:
  schemas:
...
    EvmAddress:
      type: string
      pattern: "^0x[0-9a-fA-F]{40}$"
      description: Normalized lowercase EVM address
...
    BlockNumber:
      type: string
      pattern: "^[0-9]+$"
      description: Decimal string to avoid JS precision loss
...
    IsoTimestamp:
      type: string
      format: date-time
      description: RFC 3339 UTC
    DataStatus:
      type: string
      enum:
        - MOCK_DATA
        - SYNCING
        - LIVE
        - STALE
        - DEGRADED
        - UNAVAILABLE

    # ── Envelope ──
    EnvelopeMeta:
      type: object
      properties:
        project:
          type: string
          example: "PANGU2"
        environment:
          type: string
          example: "LOCAL"
        chain_id:
          type: integer
          example: 31337
        data_status:
          $ref: "#/components/schemas/DataStatus"
        block_number:
          $ref: "#/components/schemas/BlockNumber"
          nullable: true
        generated_at:
          $ref: "#/components/schemas/IsoTimestamp"
        schema_version:
          type: string
          example: "1.0.0"
...
    SessionInfo:
      type: object
      properties:
        token:
          type: string
        wallet_address:
          $ref: "#/components/schemas/EvmAddress"
        expires_at:
          $ref: "#/components/schemas/IsoTimestamp"

    # ── Paginated List ──
...
    Envelope_Session:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SessionInfo"
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
        error:
          nullable: true
...